WP 4.5 Compatibility.
Deal with the fall-out of the patch for Trac ticket 21319 which corrects the return of `is_textdomain_loaded()` by only adding actually loaded text-domains to the `$l10n` global.

As this global was used up to now to detect whether any translation calls were made for text-domains without a `load` call, this functionality broke.

Fixed now by low level hooking into the actual translation functions for WP 4.5 and up.

@see https://core.trac.wordpress.org/ticket/21319

Closes #6